refactor: extract HealthCheckHandler to a separate file
