+++ internal/server/web_server.go
@@ -60,19 +60,6 @@ func (s *RPCServer) Shutdown() error {
 	return s.httpServer.Shutdown(context.Background())
 }
 
-type HealthCheckHandler struct {
-	logger           *zap.Logger
-	routerCollection route.RouterCollection
-}
-
-func (h *HealthCheckHandler) ServeHTTP(writer http.ResponseWriter, _ *http.Request) {
-	if h.routerCollection.IsInitialized() {
-		respondRaw(h.logger, writer, []byte("OK"), http.StatusOK)
-	} else {
-		respondRaw(h.logger, writer, []byte("Starting up"), http.StatusServiceUnavailable)
-	}
-}
-
 type RPCHandler struct {
 	router route.Router
 	logger *zap.Logger
